replace decimal.Parse with double.Parse for values greater than decimal's range; parse exclusivemax and min values as boolean types

+++ src/Microsoft.OpenApi.Readers/V2/JsonSchemaDeserializer.cs
@@ -30,13 +30,13 @@ namespace Microsoft.OpenApi.Readers.V2
             {
                 "multipleOf", (o, n) =>
                 {
-                    o.MultipleOf(decimal.Parse(n.GetScalarValue(), NumberStyles.Float, CultureInfo.InvariantCulture));
+                    o.MultipleOf(double.Parse(n.GetScalarValue(), NumberStyles.Float, CultureInfo.InvariantCulture));
                 }
             },
             {
                 "maximum", (o, n) =>
                 {
-                    o.Maximum(decimal.Parse(n.GetScalarValue(), NumberStyles.Float, CultureInfo.InvariantCulture));
+                    o.Maximum(double.Parse(n.GetScalarValue(), NumberStyles.Float, CultureInfo.InvariantCulture));
                 }
             },
             {
@@ -48,7 +48,7 @@ namespace Microsoft.OpenApi.Readers.V2
             {
                 "minimum", (o, n) =>
                 {
-                    o.Minimum(decimal.Parse(n.GetScalarValue(), NumberStyles.Float, CultureInfo.InvariantCulture));
+                    o.Minimum(double.Parse(n.GetScalarValue(), NumberStyles.Float, CultureInfo.InvariantCulture));
                 }
             },
             {
